+++ src/__generated__/Artwork_artwork.graphql.ts
@@ -2,10 +2,11 @@
 
 import { ConcreteFragment } from "relay-runtime";
 import { ArtworkActions_artwork$ref } from "./ArtworkActions_artwork.graphql";
+import { ArtworkAvailability_artwork$ref } from "./ArtworkAvailability_artwork.graphql";
 declare const _Artwork_artwork$ref: unique symbol;
 export type Artwork_artwork$ref = typeof _Artwork_artwork$ref;
 export type Artwork_artwork = {
-    readonly " $fragmentRefs": ArtworkActions_artwork$ref;
+    readonly " $fragmentRefs": ArtworkActions_artwork$ref & ArtworkAvailability_artwork$ref;
     readonly " $refType": Artwork_artwork$ref;
 };
 
@@ -24,6 +25,11 @@ const node: ConcreteFragment = {
       "args": null
     },
     {
+      "kind": "FragmentSpread",
+      "name": "ArtworkAvailability_artwork",
+      "args": null
+    },
+    {
       "kind": "ScalarField",
       "alias": null,
       "name": "__id",
@@ -32,5 +38,5 @@ const node: ConcreteFragment = {
     }
   ]
 };
-(node as any).hash = 'b14f50eb0c2c902587b938a1974cbf81';
+(node as any).hash = '1956bbb38aaf7cfafd6059f82057343f';
 export default node;