+++ storybook/storyLoader.js
@@ -41,6 +41,7 @@ function loadStories() {
   require('../src/lib/Containers/__stories__/Inquiry.story');
   require('../src/lib/Containers/__stories__/Sale.story');
   require('../src/lib/Scenes/Artwork/Components/__stories__/ArtworkActions.story');
+  require('../src/lib/Scenes/Artwork/Components/__stories__/ArtworkAvailability.story');
   require('../src/lib/Scenes/Artwork/__stories__/Artwork.story');
   require('../src/lib/Scenes/Fair/__stories__/Fair.story');
   require('../src/lib/Scenes/Favorites/__stories__/Favorites.story');
@@ -89,6 +90,7 @@ const stories = [
   '../src/lib/Containers/__stories__/Inquiry.story',
   '../src/lib/Containers/__stories__/Sale.story',
   '../src/lib/Scenes/Artwork/Components/__stories__/ArtworkActions.story',
+  '../src/lib/Scenes/Artwork/Components/__stories__/ArtworkAvailability.story',
   '../src/lib/Scenes/Artwork/__stories__/Artwork.story',
   '../src/lib/Scenes/Fair/__stories__/Fair.story',
   '../src/lib/Scenes/Favorites/__stories__/Favorites.story',